Refactor DTOs into feature-based folders and files
Split Category, TimeLog, and User DTOs into individual files under feature-specific namespaces. Updated TimelogAPI.csproj to remove old DTO folder references. Improves code modularity and maintainability.
