--- TimelogAPI/DTOs/TimeLogDTO.cs
@@ -1,25 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace TimelogAPI.DTOs
-{
-    public record TimeLogResponse(
-        int Id,
-        DateTime StartTime,
-        DateTime? EndTime,
-        int CategoryId
-    );
-
-    public record CreateTimeLogRequest(
-        [Required]
-        DateTime StartTime,
-        DateTime? EndTime,
-        [Required]
-        int CategoryId
-    );
-
-    public record UpdateTimeLogRequest(
-        DateTime? StartTime,
-        DateTime? EndTime,
-        int? CategoryId
-        );
-}
