+++ TimelogAPI/Features/Category/Dtos/CategoryResponse.cs
@@ -1,0 +1,10 @@
+﻿using TimelogAPI.Features.TimeLog.Dtos;
+
+namespace TimelogAPI.Features.Category.Dtos
+{
+    public record CategoryResponse(
+        int Id,
+        string Name,
+        List<TimeLogResponse> TimeLogs
+    );
+}
