+++ TimelogAPI/Features/Category/Dtos/UpdateCategoryRequest.cs
@@ -1,0 +1,6 @@
+﻿namespace TimelogAPI.Features.Category.Dtos
+{
+    public record UpdateCategoryRequest(
+        string Name
+    );
+}
