+++ TimelogAPI/Features/TimeLog/Dtos/TimeLogResponse.cs
@@ -1,0 +1,9 @@
+﻿namespace TimelogAPI.Features.TimeLog.Dtos
+{
+    public record TimeLogResponse(
+        int Id,
+        DateTime StartTime,
+        DateTime? EndTime,
+        int CategoryId
+    );
+}
