+++ TimelogAPI/Features/TimeLog/Dtos/UpdateTimeLogRequest.cs
@@ -1,0 +1,8 @@
+﻿namespace TimelogAPI.Features.TimeLog.Dtos
+{
+    public record UpdateTimeLogRequest(
+        DateTime? StartTime,
+        DateTime? EndTime,
+        int? CategoryId
+    );
+}
