+++ TimelogAPI/TimelogAPI.csproj
@@ -13,8 +13,6 @@
 
   <ItemGroup>
     <Folder Include="Controllers\" />
-    <Folder Include="Features\Category\Dtos\" />
-    <Folder Include="Features\TimeLog\Dtos\" />
   </ItemGroup>
 
 </Project>
